+++ packages/houdini-svelte/src/plugin/codegen/stores/fragment.ts
@@ -2,7 +2,7 @@ import type { Document, GenerateHookInput } from 'houdini'
 import { fs, path } from 'houdini'
 
 import type { HoudiniSvelteConfig } from '../../config'
-import { stores_directory, store_name } from '../../kit'
+import { stores_directory, store_name } from '../../storeConfig'
 import { store_import } from './custom'
 
 export async function fragmentStore({ config, pluginRoot }: GenerateHookInput, doc: Document) {
