+++ packages/houdini-svelte/src/plugin/codegen/stores/query.ts
@@ -3,7 +3,7 @@ import type { Document, GenerateHookInput } from 'houdini'
 import { fs, operation_requires_variables, path } from 'houdini'
 
 import type { HoudiniSvelteConfig } from '../../config'
-import { stores_directory, store_name } from '../../kit'
+import { stores_directory, store_name } from '../../storeConfig'
 import { store_import } from './custom'
 
 export async function queryStore({ config, pluginRoot }: GenerateHookInput, doc: Document) {
