+++ packages/houdini-svelte/src/plugin/extractLoadFunction.ts
@@ -6,7 +6,7 @@ import { fs, parseJS, path } from 'houdini'
 import type * as recast from 'recast'
 import { transformWithEsbuild } from 'vite'
 
-import { stores_directory_name, store_suffix } from './kit'
+import { stores_directory_name, store_suffix } from './storeConfig'
 import { houdini_load_fn } from './naming'
 
 type Program = recast.types.namedTypes.Program
